Simplify TLS deserialization.
The TLSDeserializer class contained a mix of low-level TLS deserializing
and deserialization of application-specific data (SCTs, etc).

The TLSDeserializer was moved to tls_encoding.h/cc and was left only with
the basic TLS deserialization abilities.

De-serialization of application-specific structures remains in serializer.h
as stand-alone functions (rather than member functions).
